requirements.txt
Django==1.11.9
djangorestframework==3.8.2
pytz==2018.5
